+++ linux/flutter/generated_plugins.cmake
@@ -7,6 +7,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   dynamic_color
   open_file_linux
   screen_retriever_linux
+  url_launcher_linux
   window_manager
 )