+++ windows/flutter/generated_plugins.cmake
@@ -7,6 +7,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   dynamic_color
   permission_handler_windows
   screen_retriever_windows
+  url_launcher_windows
   window_manager
 )